Convert radial function tests to double precision

+++ test/cpp/math/radial_function.test.hpp
@@ -16,7 +16,7 @@ using namespace std::string_literals;
 TEST_CASE("Testing radial function") {
   RadialFunction R;
 
-  for (float r = 0; r <= 4; r += 0.25) {
+  for (double r = 0; r <= 4; r += 0.25) {
     R.setR(BOHR_TO_PM(r));
     SUBCASE(("r: "s + std::to_string(r) + " bohrs").c_str()) {
       SUBCASE("n: 1") {
@@ -24,7 +24,7 @@ TEST_CASE("Testing radial function") {
 
         SUBCASE("l: 0") {
           R.setL(0);
-          const float expected = 2.f * expf(-r);
+          const double expected = 2. * exp(-r);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
       }
@@ -34,13 +34,13 @@ TEST_CASE("Testing radial function") {
 
         SUBCASE("l: 0") {
           R.setL(0);
-          const float expected = (1 / sqrtf(2)) * (1 - (r / 2)) * expf(-r / 2);
+          const double expected = (1 / sqrt(2)) * (1 - (r / 2)) * exp(-r / 2);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
 
         SUBCASE("l: 1") {
           R.setL(1);
-          const float expected = (1 / sqrtf(24)) * r * expf(-r / 2);
+          const double expected = (1 / sqrt(24)) * r * exp(-r / 2);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
       }
@@ -50,20 +50,20 @@ TEST_CASE("Testing radial function") {
 
         SUBCASE("l: 0") {
           R.setL(0);
-          const float expected =
-              (2 / sqrtf(27)) * (1 - ((2 * r) / 3) + ((2 * powf(r, 2)) / 27)) * expf(-r / 3);
+          const double expected =
+              (2 / sqrt(27)) * (1 - ((2 * r) / 3) + ((2 * pow(r, 2)) / 27)) * exp(-r / 3);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
 
         SUBCASE("l: 1") {
           R.setL(1);
-          const float expected = (8 / (27 * sqrtf(6))) * r * (1 - (r / 6)) * expf(-r / 3);
+          const double expected = (8 / (27 * sqrt(6))) * r * (1 - (r / 6)) * exp(-r / 3);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
 
         SUBCASE("l: 2") {
           R.setL(2);
-          const float expected = (4 / (81 * sqrtf(30))) * powf(r, 2) * expf(-r / 3);
+          const double expected = (4 / (81 * sqrt(30))) * pow(r, 2) * exp(-r / 3);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
       }
@@ -71,31 +71,31 @@ TEST_CASE("Testing radial function") {
       SUBCASE("n: 4") {
         R.setN(4);
 
-        const float roh = r / 2;
+        const double roh = r / 2;
 
         SUBCASE("l: 0") {
           R.setL(0);
-          const float expected =
-              (1.f / 96.f) * (24 - 36 * roh + 12 * powf(roh, 2) - powf(roh, 3)) * expf(-r / 4);
+          const double expected =
+              (1. / 96.) * (24 - 36 * roh + 12 * pow(roh, 2) - pow(roh, 3)) * exp(-r / 4);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
 
         SUBCASE("l: 1") {
           R.setL(1);
-          const float expected =
-              (1 / (32 * sqrtf(15))) * (20 - 10 * roh + powf(roh, 2)) * roh * expf(-r / 4);
+          const double expected =
+              (1 / (32 * sqrt(15))) * (20 - 10 * roh + pow(roh, 2)) * roh * exp(-r / 4);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
 
         SUBCASE("l: 2") {
           R.setL(2);
-          const float expected = (1 / (96 * sqrtf(5))) * (6 - roh) * powf(roh, 2) * expf(-r / 4);
+          const double expected = (1 / (96 * sqrt(5))) * (6 - roh) * pow(roh, 2) * exp(-r / 4);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
 
         SUBCASE("l: 3") {
           R.setL(3);
-          const float expected = (1 / (96 * sqrtf(35))) * powf(roh, 3) * expf(-r / 4);
+          const double expected = (1 / (96 * sqrt(35))) * pow(roh, 3) * exp(-r / 4);
           CHECK(R() == doctest::Approx(expected).epsilon(.000001));
         }
       }
